requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic-settings
python-dotenv
azure-identity
opencensus-ext-azure